src/python/workshop/mcp/requirements.txt
mcp[cli]>=1.4.1
azure-identity>=1.21.0
python-dotenv>=1.0.1
azure-ai-projects==1.0.0b8